refactor(workspace): extract teamclaw-p2p crate from monolithic src-tauri
Move all iroh-based P2P logic (team sync, member management, gossip,
namespace rotation) into crates/teamclaw-p2p. The main crate now holds
thin Tauri command wrappers that delegate to the new crate and supply
app-level constants (TEAMCLAW_DIR, CONFIG_FILE_NAME).

Also cfg-gate iroh SecretKey usage in oss_commands::get_device_id so
the non-p2p (Windows) build path compiles without iroh.
