+++ src-tauri/src/commands/team_unified.rs
@@ -110,7 +110,7 @@ async fn p2p_update_member_role_and_sync(
         })
         .flatten()
         .or_else(|| {
-            super::team_p2p::read_p2p_config(workspace_path)
+            super::team_p2p::read_p2p_config(workspace_path, super::TEAMCLAW_DIR, super::CONFIG_FILE_NAME)
                 .ok()
                 .flatten()
                 .and_then(|config| {
@@ -123,7 +123,7 @@ async fn p2p_update_member_role_and_sync(
         })
         .ok_or_else(|| "Member not found".to_string())?;
 
-    super::team_p2p::update_member_role(workspace_path, &team_dir, &caller_node_id, node_id, role)?;
+    super::team_p2p::update_member_role(workspace_path, &team_dir, &caller_node_id, node_id, role, super::TEAMCLAW_DIR, super::CONFIG_FILE_NAME)?;
 
     let manifest_path = format!("{}/{}", team_dir, "_team/members.json");
     let content = match std::fs::read(&manifest_path) {
@@ -135,6 +135,8 @@ async fn p2p_update_member_role_and_sync(
                 &caller_node_id,
                 node_id,
                 previous_role,
+                super::TEAMCLAW_DIR,
+                super::CONFIG_FILE_NAME,
             );
             return Err(format!(
                 "Could not read members.json after role update (local role restored): {}",
@@ -154,6 +156,8 @@ async fn p2p_update_member_role_and_sync(
             &caller_node_id,
             node_id,
             previous_role,
+            super::TEAMCLAW_DIR,
+            super::CONFIG_FILE_NAME,
         );
         return Err(format!(
             "Failed to sync role change to the team document (local role restored): {}. Reconnect and try again.",
@@ -196,7 +200,7 @@ pub async fn unified_team_get_members(
             if let Ok(Some(manifest)) = super::team_p2p::read_members_manifest(&team_dir) {
                 Ok(manifest.members)
             } else {
-                let config = super::team_p2p::read_p2p_config(&workspace_path)?.unwrap_or_default();
+                let config = super::team_p2p::read_p2p_config(&workspace_path, super::TEAMCLAW_DIR, super::CONFIG_FILE_NAME)?.unwrap_or_default();
                 Ok(config.allowed_members)
             }
         }
@@ -253,6 +257,8 @@ pub async fn unified_team_add_member(
                 &team_dir,
                 &caller_node_id,
                 member,
+                super::TEAMCLAW_DIR,
+                super::CONFIG_FILE_NAME,
             )?;
             // Push updated members.json to Iroh doc immediately (don't rely on fs watcher).
             // Joiners authorize against the doc, not only local disk.
@@ -265,6 +271,8 @@ pub async fn unified_team_add_member(
                         &team_dir,
                         &caller_node_id,
                         &added_node_id,
+                        super::TEAMCLAW_DIR,
+                        super::CONFIG_FILE_NAME,
                     );
                     return Err(format!(
                         "Could not read members.json after add (rolled back): {}",
@@ -282,6 +290,8 @@ pub async fn unified_team_add_member(
                     &team_dir,
                     &caller_node_id,
                     &added_node_id,
+                    super::TEAMCLAW_DIR,
+                    super::CONFIG_FILE_NAME,
                 );
                 return Err(format!(
                     "Failed to sync members to the team document (rolled back): {}. Reconnect and try again.",
@@ -338,7 +348,7 @@ pub async fn unified_team_remove_member(
                         .to_string(),
                 );
             }
-            let member_snapshot = super::team_p2p::read_p2p_config(&workspace_path)?
+            let member_snapshot = super::team_p2p::read_p2p_config(&workspace_path, super::TEAMCLAW_DIR, super::CONFIG_FILE_NAME)?
                 .ok_or_else(|| "No P2P config found".to_string())?
                 .allowed_members
                 .iter()
@@ -351,6 +361,8 @@ pub async fn unified_team_remove_member(
                 &team_dir,
                 &caller_node_id,
                 &node_id,
+                super::TEAMCLAW_DIR,
+                super::CONFIG_FILE_NAME,
             )?;
 
             let manifest_path = format!("{}/{}", team_dir, "_team/members.json");
@@ -362,6 +374,8 @@ pub async fn unified_team_remove_member(
                         &team_dir,
                         &caller_node_id,
                         member_snapshot.clone(),
+                        super::TEAMCLAW_DIR,
+                        super::CONFIG_FILE_NAME,
                     ) {
                         return Err(format!(
                             "Could not read members.json after remove (restore failed): read {}, revert {}",
@@ -385,6 +399,8 @@ pub async fn unified_team_remove_member(
                     &team_dir,
                     &caller_node_id,
                     member_snapshot,
+                    super::TEAMCLAW_DIR,
+                    super::CONFIG_FILE_NAME,
                 ) {
                     return Err(format!(
                         "Failed to sync removal to the team document: {}. Also failed to restore local member list: {}.",
@@ -474,7 +490,7 @@ pub async fn unified_team_get_my_role(
         Some("p2p") => {
             let _ = iroh_state;
             let config =
-                super::team_p2p::read_p2p_config(&workspace_path)?.ok_or("No P2P config found")?;
+                super::team_p2p::read_p2p_config(&workspace_path, super::TEAMCLAW_DIR, super::CONFIG_FILE_NAME)?.ok_or("No P2P config found")?;
             config
                 .role
                 .ok_or_else(|| "Role not set in P2P config".to_string())
